issues/9 refactor(rename): ICacheLocator -> IUrlLocator
