+++ Assets/Editor/Test/FileCacheTest.cs
@@ -1,5 +1,6 @@
 ﻿using System.IO;
 using NUnit.Framework;
+using Unicache.Plugin;
 using UniRx;
 
 namespace Unicache.Test
@@ -66,7 +67,7 @@ namespace Unicache.Test
         [Test]
         public void FetchTest()
         {
-            this.cache.Locator = new TestCacheLocator();
+            this.cache.UrlLocator = new SimpleUrlLocator();
             this.cache.Handler = new TestCacheHandler();
 
             int count = 0;
@@ -98,14 +99,6 @@ namespace Unicache.Test
             protected override IObservable<byte[]> AsAsync(IObservable<byte[]> observable)
             {
                 return observable;
-            }
-        }
-
-        class TestCacheLocator : ICacheLocator
-        {
-            public string CreatePath(string url)
-            {
-                return url;
             }
         }
