--- Assets/Plugins/Unicache/Scripts/ICacheLocator.cs
@@ -1,8 +1,0 @@
-﻿namespace Unicache
-{
-    // CacheLocator determinates save path from resource url
-    public interface ICacheLocator
-    {
-        string CreatePath(string url);
-    }
-}
