+++ Assets/Plugins/Unicache/Scripts/IUnicache.cs
@@ -5,13 +5,13 @@ namespace Unicache
     public interface IUnicache
     {
         ICacheHandler Handler { set; }
-        ICacheLocator Locator { set; }
+        IUrlLocator UrlLocator { set; }
 
-        IObservable<byte[]> Fetch(string url);
+        IObservable<byte[]> Fetch(string key);
         void ClearAll();
 
-        byte[] GetCache(string path /*, string hash */);
-        void SetCache(string path /*, string hash*/, byte[] data);
-        bool HasCache(string path /*, string hash*/);
+        byte[] GetCache(string path);
+        void SetCache(string path, byte[] data);
+        bool HasCache(string path);
     }
 }
