+++ Assets/Plugins/Unicache/Scripts/IUrlLocator.cs
@@ -1,0 +1,7 @@
+﻿namespace Unicache
+{
+    public interface IUrlLocator
+    {
+        string CreateUrl(string key);
+    }
+}
