--- Assets/Plugins/Unicache/Scripts/Plugin/SimpleCacheLocator.cs
@@ -1,10 +1,0 @@
-namespace Unicache.Plugin
-{
-    public class SimpleCacheLocator : ICacheLocator
-    {
-        public string CreatePath(string url)
-        {
-            return Digest.SHA1(url);
-        }
-    }
-}
