+++ Assets/Plugins/Unicache/Scripts/Plugin/SimpleUrlLocator.cs
@@ -1,0 +1,10 @@
+namespace Unicache.Plugin
+{
+    public class SimpleUrlLocator : IUrlLocator
+    {
+        public string CreateUrl(string key)
+        {
+            return key;
+        }
+    }
+}
